+++ sparkles-parser/src/tracing_decoder.rs
@@ -29,8 +29,10 @@ pub enum ParsingState {
     /// id, dif_tm_len
     DifTm(TracingEventId, usize),
 
-    RangeOrdId(Option<TracingEventId>, usize),
-    RangeTm(Option<TracingEventId>, usize, u8)
+    RangeOrdId(Option<TracingEventId>, usize, bool),
+    RangeTm(Option<TracingEventId>, usize, u8, bool),
+    ForeignThreadIdLen(Option<TracingEventId>, u64, u8),
+    ForeignThreadId(Option<TracingEventId>, u64, u8, usize)
 }
 
 impl StreamFrameDecoder {
@@ -47,14 +49,15 @@ impl StreamFrameDecoder {
 
                 let is_range_event = dif_tm_len & 0b1000_0000 != 0;
                 let is_unnamed_range_end = dif_tm_len & 0b0100_0000 != 0;
-                let dif_tm_len = (dif_tm_len & 0b0000_1111) as usize;
+                let has_foreign_thread = dif_tm_len & 0b0010_0000 != 0;
+                let dif_tm_len = (dif_tm_len & 0b0001_1111) as usize;
 
                 if is_range_event {
                     if is_unnamed_range_end {
-                        (None, ParsingState::RangeOrdId(None, dif_tm_len))
+                        (None, ParsingState::RangeOrdId(None, dif_tm_len, has_foreign_thread))
                     }
                     else {
-                        (None, ParsingState::RangeOrdId(Some(ev), dif_tm_len))
+                        (None, ParsingState::RangeOrdId(Some(ev), dif_tm_len, has_foreign_thread))
                     }
                 }
                 else {
@@ -67,22 +70,38 @@ impl StreamFrameDecoder {
                 let dif_tm = u64::from_le_bytes(buf);
                 (Some(TracingEvent::Instant(ev, dif_tm)), ParsingState::NewFrame)
             }
-            ParsingState::RangeOrdId(ev, dif_tm_len) if available_bytes_len >= 1 => {
+            ParsingState::RangeOrdId(ev, dif_tm_len, has_foreign_thread) if available_bytes_len >= 1 => {
                 let ord_id = self.buf.try_pop().unwrap();
 
-                (None, ParsingState::RangeTm(ev, dif_tm_len, ord_id))
+                (None, ParsingState::RangeTm(ev, dif_tm_len, ord_id, has_foreign_thread))
             }
-            ParsingState::RangeTm(ev_id, dif_tm_len, ord_id) if available_bytes_len >= dif_tm_len => {
+            ParsingState::RangeTm(ev_id, dif_tm_len, ord_id, has_foreign_thread) if available_bytes_len >= dif_tm_len => {
                 let mut buf = [0u8; 8];
                 self.buf.pop_slice(&mut buf[..dif_tm_len]);
                 let dif_tm = u64::from_le_bytes(buf);
 
-                let ev = if let Some(id) = ev_id {
-                    Some(TracingEvent::RangePart(id, dif_tm, ord_id))
+                if has_foreign_thread {
+                    (None, ParsingState::ForeignThreadIdLen(ev_id, dif_tm, ord_id))
+                } else {
+                    let ev = if let Some(id) = ev_id {
+                        Some(TracingEvent::RangePart(id, dif_tm, ord_id))
+                    }
+                    else {
+                        Some(TracingEvent::UnnamedRangeEnd(dif_tm, ord_id))
+                    };
+                    (ev, ParsingState::NewFrame)
                 }
-                else {
-                    Some(TracingEvent::UnnamedRangeEnd(dif_tm, ord_id))
-                };
+            }
+            ParsingState::ForeignThreadIdLen(ev_id, dif_tm, ord_id) if available_bytes_len >= 1 => {
+                let foreign_thread_id_bytes_len = self.buf.try_pop().unwrap() as usize;
+                (None, ParsingState::ForeignThreadId(ev_id, dif_tm, ord_id, foreign_thread_id_bytes_len))
+            }
+            ParsingState::ForeignThreadId(ev_id, dif_tm, ord_id, foreign_thread_id_bytes_len) if available_bytes_len >= foreign_thread_id_bytes_len => {
+                let mut buf = [0u8; 8];
+                self.buf.pop_slice(&mut buf[..foreign_thread_id_bytes_len]);
+                let foreign_thread_id = u64::from_le_bytes(buf);
+
+                let ev = Some(TracingEvent::ForeignRangeEnd(ev_id, dif_tm, ord_id, foreign_thread_id));
                 (ev, ParsingState::NewFrame)
             }
             state => {