+++ CHANGELOG.md
@@ -14,15 +14,24 @@ For protocol used by both crates there are some guarantees based on version numb
 ### [Unreleased] sparkles
 - Add `unstable-thread-id` feature for platforms without `thread-id` implementation (requires nightly)
 - Fix panic on systems without process_name
-- Examples: Enable UDP multicast by default
+- Config: rename `flush_threshold` to `sending_threshold`
+- Config: add `auto_send_ms` for configuring maximum interval for sending trace data packets, improving responsiveness in real-time parsing scenarios.
+Examples:
+- Enable UDP multicast by default
+- Add `cross-thread-ranges` example
 
 ### [Unreleased] sparkles-core
 - Add RISC-V 32-bit architecture timestamp support
-- Improve target detection
+- Refactor target detection
 - Add `force-fallback-impl` feature to use `Instant`-based implementation instead of arch-specific
+- Add timeout flush trigger, improving responsiveness in real-time parsing scenarios.
+- Config: add `auto_flush_ms` for configuring maximum interval for flush trigger.
+- Range guard is now Send: You can start and end range event in different threads.
+- Put more stability in `IdMappingState::insert_and_get_id`: return 255 in case of overflow instead of returning wrong ids.
 
 ### [Unreleased] sparkles-parser
 - `SparklesParser` now will send arrays of parsed events to the callback instead of single event.
+- Implement parsing for new cross-thread events type.
 
 ## Versions
 ### [0.1.8] sparkles (proto-1.0)